CHANGE: New relations, between CE and tags, not between sites and tags
The base relation is between CE-Tag because the Sites can contain
multiple CEs that are not necessarily using the same software directory. 
When the tag is registered at a site, we assume that ALL CEs will have
access to it. Then we need to get jobs running there to validate this.
In principle, we want to setup probe jobs that validate the status
automatically. NEED more tag statuses!

Ideally the jobs should target only the CEs that are correct. This is
not yet possible but relies on the RSS. Several questions are related to
that, and we need to make sure that the RSS will be able to deal with
it.